Correct description of Log4j 1 to Log4j 2 APi migration
The description of the "Migrate Log4j 1.x to Log4j 2.x" recipe is incorrect.

Users might get the false impression that it mitigates the Log4Shell vulnerability. However:

- Log4j 1 was **never** affected by Log4Shell.
- only Log4j 2 Core was.

There are multiple reasons to migrate from Log4j 1, the first one is that the library is unsupported since 2015. Log4Shell mitigation is **not** one of them.