`Slf4jToLog4j`: No import for nested `CloseableThreadContext.Instance` type
In the declarative recipes fully qualified names are often not specified in binary form (with `$` separators for nested types) and as a consequence it is difficult to determine what part of the qualified name forms the package (although even with `$` in the name we just assume that this is a nested type separator).

By specifying the name in binary form (`org.apache.logging.log4j.CloseableThreadContext$Instance`) the `ChangeType` recipe behaves as expected by not adding an import for the nested type.